moving controllers to scripts folder
