+++ scripts/addStudentCourses.js
