+++ scripts/getAllCoursePlans.js
