+++ scripts/getDegreeProgress.js
