+++ scripts/getEligibleCourses.js
