+++ scripts/getPlannedCourses.js
