+++ scripts/getStudentCourses.js
