+++ scripts/updateCoursePlan.js
